requirements.txt
kgisuperpy>=1.0.2
